+++ src/pydocformatter/rules/definitions/PCF/PCF001_standalone_comment_formatting.py
@@ -35,6 +35,7 @@ class PCF001StandaloneCommentFormatting(RuleBase):
         fix_availability=FixAvailability.ALWAYS,
         stable_since="0.3.0",
         setting_effects=(),
+        incompatible_with=(),
     )
 
     @classmethod