+++ src/pydocformatter/rules/definitions/PCF/PCF002_trailing_comment_formatting.py
@@ -19,6 +19,7 @@ class PCF002TrailingCommentFormatting(RuleBase):
         fix_availability=FixAvailability.ALWAYS,
         stable_since="0.3.0",
         setting_effects=(),
+        incompatible_with=(),
     )
 
     @classmethod