+++ src/pydocformatter/rules/definitions/PDF/PDF000_concatenated_docstring_literal.py
@@ -17,6 +17,7 @@ class PDF000ConcatenatedDocstringLiteral(RuleBase):
         fix_availability=FixAvailability.ALWAYS,
         stable_since="0.3.0",
         setting_effects=(),
+        incompatible_with=(),
     )
 
     @classmethod