+++ src/pydocformatter/rules/definitions/PDF/PDF004_docstring_blank_line_whitespace.py
@@ -16,4 +16,5 @@ class PDF004DocstringBlankLineWhitespace(RuleBase):
         fix_availability=FixAvailability.ALWAYS,
         stable_since="0.3.0",
         setting_effects=(),
+        incompatible_with=(),
     )