+++ src/pydocformatter/rules/definitions/PDF/PDF101_multiline_opening_quotes_same_line.py
@@ -22,4 +22,5 @@ class PDF101MultilineOpeningQuotesSameLine(RuleBase):
                 effects=(RuleSettingEffectValues(effect=RuleSettingEffect.IGNORED, values=(DocstringConvention.NUMPY, DocstringConvention.PEP257)),),
             ),
         ),
+        incompatible_with=(RuleCode("PDF102"),),
     )