+++ src/pydocformatter/rules/definitions/PDF/PDF106_docstring_should_be_one_line.py
@@ -16,4 +16,5 @@ class PDF106DocstringShouldBeOneLine(RuleBase):
         fix_availability=FixAvailability.ALWAYS,
         stable_since="0.3.0",
         setting_effects=(),
+        incompatible_with=(),
     )